added child item handling to modifier choice prompt
