+++ src/module/actor/base/sheet.ts
@@ -42,8 +42,6 @@ export class ActorSheetGURPS extends ActorSheet {
 		event: DragEvent,
 		data: ActorSheet.DropData.Item & { actor: BaseActorGURPS; _uuid?: string }
 	): Promise<unknown> {
-		// Const element = $(event.currentTarget!)
-		// const widthAcross = (event.pageX! - element.offset()!.left) / element.width()!
 		const top = Boolean($(".border-top").length)
 		const inContainer = Boolean($(".border-in").length)
 
@@ -55,7 +53,6 @@ export class ActorSheetGURPS extends ActorSheet {
 
 		if (!this.actor.isOwner) return false
 
-		// Const item = await (BaseItemGURPS as any).implementation.fromDropData(data);
 		let item: Item
 		if (data._uuid) {
 			const importData = {
@@ -69,7 +66,7 @@ export class ActorSheetGURPS extends ActorSheet {
 		const itemData = { ...item.toObject(), uuid: item.uuid }
 
 		// Handle item sorting within the same Actor
-		console.log(itemData, top, inContainer)
+		// console.log(itemData, top, inContainer)
 		if (this.actor.uuid === item.actor?.uuid) {
 			console.log(top, inContainer)
 			return this._onSortItem(event, itemData, { top: top, in: inContainer })
@@ -130,10 +127,8 @@ export class ActorSheetGURPS extends ActorSheet {
 		itemData: PropertiesToSource<ItemDataBaseProperties> & { uuid: string },
 		options: { top: boolean; in: boolean } = { top: false, in: false }
 	): Promise<Item[]> {
-		// Const source: any = this.actor.deepItems.get(itemData._id!)
 		const source: any = this.actor.deepItems.get(itemData.uuid)
 		let dropTarget = $(event.target!).closest(".desc[data-uuid]")
-		// Let target: any = this.actor.deepItems.get(dropTarget?.data("uuid")?.split(".").at(-1))
 		let target: any = this.actor.deepItems.get(dropTarget?.data("uuid"))
 		if (!target) return []
 		let parent: any = target?.parent
