+++ src/module/item/container/sheet.ts
@@ -28,20 +28,12 @@ export class ContainerSheetGURPS extends ItemSheetGURPS {
 		super.activateListeners(html)
 		html.find(".dropdown-toggle").on("click", event => this._onCollapseToggle(event))
 		html.find(".enabled").on("click", event => this._onEnabledToggle(event))
-		// Html.find(".item-list").on("dragend", event => this._onDrop(event));
 	}
 
 	protected override async _onDragStart(event: DragEvent): Promise<void> {
 		const list = event.currentTarget
-		// If (event.target.classList.contains("contents-link")) return;
 
 		let dragData: any
-		// Const dragData: any = {
-		// 	actorId: this.actor.id,
-		// 	sceneId: this.actor.isToken ? canvas?.scene?.id : null,
-		// 	tokenId: this.actor.isToken ? this.actor.token?.id : null,
-		// 	pack: this.actor?.pack,
-		// };
 
 		// Owned Items
 		if ((list as HTMLElement).dataset.itemId) {
@@ -104,8 +96,12 @@ export class ContainerSheetGURPS extends ItemSheetGURPS {
 
 	// DragData handling
 	protected async _onDropItem(event: DragEvent, data: ActorSheet.DropData.Item): Promise<unknown> {
-		// Remove Drag Markers
-		$(".drop-over").removeClass("drop-over")
+		const top = Boolean($(".border-top").length)
+		const inContainer = Boolean($(".border-in").length)
+
+		$(".border-bottom").removeClass("border-bottom")
+		$(".border-top").removeClass("border-top")
+		$(".border-in").removeClass("border-in")
 
 		if (!this.item.isOwner) return false
 
@@ -114,7 +110,8 @@ export class ContainerSheetGURPS extends ItemSheetGURPS {
 		const itemData = { ...item.toObject(), uuid: item.uuid }
 
 		// Handle item sorting within the same Actor
-		if (this.item.uuid === item.parent?.uuid) return this._onSortItem(event, itemData)
+		if (this.item.uuid === item.parent?.uuid)
+			return this._onSortItem(event, itemData, { top: top, in: inContainer })
 
 		return this._onDropItemCreate(itemData)
 	}
@@ -126,30 +123,40 @@ export class ContainerSheetGURPS extends ItemSheetGURPS {
 
 	protected async _onSortItem(
 		event: DragEvent,
-		itemData: PropertiesToSource<ItemDataBaseProperties> & { uuid: string }
+		itemData: PropertiesToSource<ItemDataBaseProperties> & { uuid: string },
+		options: { top: boolean; in: boolean } = { top: false, in: false }
 	): Promise<Item[]> {
-		// Const source = (this.item as any).deepItems.get(itemData._id!)
-		const source = (this.item as any).deepItems.get(itemData.uuid)
-		const dropTarget = $(event.target!).closest("[data-item-id]")
-		const target = (this.item as any).deepItems.get(dropTarget.data("item-id"))
+		const source: any = this.object.deepItems.get(itemData.uuid)
+		let dropTarget = $(event.target!).closest(".desc[data-uuid]")
+		let target: any = this.object.deepItems.get(dropTarget?.data("uuid"))
 		if (!target) return []
-		const parent = target?.parent as Actor | Item
-		const siblings = target!.parent!.items.filter((i: Item) => i.id !== source!.id)
+		let parent: any = target?.parent
+		let parents = target?.parents
+		if (options.in) {
+			parent = target
+			target = parent.children.contents[0] ?? null
+		}
+		const siblings = (parent!.items as Collection<Item>).filter(
+			i => i.id !== source!.id && (source as any)!.sameSection(i)
+		)
+		if (target && !(source as any)?.sameSection(target)) return []
 
 		const sortUpdates = SortingHelpers.performIntegerSort(source, {
 			target: target,
-			siblings,
+			siblings: siblings,
+			sortBefore: options.top,
 		})
 		const updateData = sortUpdates.map(u => {
 			const update = u.update
-			;(update as any)._id = u.target!.id
+			;(update as any)._id = u.target!._id
 			return update
 		})
 
-		if (source && target && source.parent !== target.parent) {
-			if (source instanceof ContainerGURPS && target.parents.includes(source)) return []
-			await source!.parent!.deleteEmbeddedDocuments("Item", [source!.id!], { render: false })
-			return (parent as any)?.createEmbeddedDocuments(
+		if (source && source.parent !== parent) {
+			if (source.items && parents.includes(source)) return []
+			console.log(source.name, "going in", parent.name)
+			await source.parent!.deleteEmbeddedDocuments("Item", [source!._id!], { render: false })
+			return parent?.createEmbeddedDocuments(
 				"Item",
 				[
 					{
@@ -184,4 +191,8 @@ export class ContainerSheetGURPS extends ItemSheetGURPS {
 		})
 		return this.render()
 	}
+}
+
+export interface ContainerSheetGURPS extends ItemSheetGURPS {
+	object: ContainerGURPS
 }
