+++ src/styles/app/compendium_browser/_embeds.scss
@@ -97,15 +97,6 @@
 		&.quantity {
 			min-width: 1em; //possibly unnecessary
 		}
-		// &.drop-over {
-		// 	&.desc {
-		// 		border-top: none;
-		// 		.name {
-		// 			border-top: 2px dashed rgb(var(--color-drop-area));
-		// 		}
-		// 	}
-		// 	border-top: 2px dashed rgb(var(--color-drop-area));
-		// }
 	}
 
 	.desc.header {
