+++ src/templates/item/sub-sheet.hbs
@@ -9,6 +9,6 @@
 	</div>
 	<div class="buttons">
 		<button id="apply">{{localize "gurps.item.substitution.apply"}}</button>
-		<button id="canel">{{localize "gurps.item.substitution.cancel"}}</button>
+		<button id="cancel">{{localize "gurps.item.substitution.cancel"}}</button>
 	</div>
 </form>
